+++ pkg/controllers/componentInstallManager.go
@@ -3,22 +3,27 @@ package controllers
 import (
 	"context"
 	"fmt"
-	"github.com/cloudogu/k8s-component-operator/pkg/api/ecosystem"
+
 	k8sv1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 // componentInstallManager is a central unit in the process of handling the installation process of a custom dogu resource.
 type componentInstallManager struct {
-	componentClient ecosystem.ComponentInterface
-	helmClient      HelmClient
+	componentClient componentInterface
+	helmClient      helmClient
+	recorder        record.EventRecorder
 }
 
 // NewComponentInstallManager creates a new instance of componentInstallManager.
-func NewComponentInstallManager(componentClient ecosystem.ComponentInterface, helmClient HelmClient) *componentInstallManager {
+func NewComponentInstallManager(componentClient componentInterface, helmClient helmClient, recorder record.EventRecorder) *componentInstallManager {
 	return &componentInstallManager{
 		componentClient: componentClient,
 		helmClient:      helmClient,
+		recorder:        recorder,
 	}
 }
 
@@ -27,9 +32,15 @@ func NewComponentInstallManager(componentClient ecosystem.ComponentInterface, he
 func (cim *componentInstallManager) Install(ctx context.Context, component *k8sv1.Component) error {
 	logger := log.FromContext(ctx)
 
-	component, err := cim.componentClient.UpdateStatusInstalling(ctx, component)
+	err := cim.helmClient.SatisfiesDependencies(ctx, component.GetHelmChartSpec())
 	if err != nil {
-		return fmt.Errorf("failed to set status installing: %w", err)
+		cim.recorder.Eventf(component, corev1.EventTypeWarning, InstallEventReason, "Dependency check failed: %s", err.Error())
+		return &genericRequeueableError{errMsg: "failed to check dependencies", err: err}
+	}
+
+	component, err = cim.componentClient.UpdateStatusInstalling(ctx, component)
+	if err != nil {
+		return &genericRequeueableError{errMsg: "failed to set status installing", err: err}
 	}
 
 	// Set the finalizer at the beginning of the installation procedure.
@@ -38,21 +49,21 @@ func (cim *componentInstallManager) Install(ctx context.Context, component *k8sv
 	// deletion procedure from the controller.
 	component, err = cim.componentClient.AddFinalizer(ctx, component, k8sv1.FinalizerName)
 	if err != nil {
-		return fmt.Errorf("failed to add finalizer %s: %w", k8sv1.FinalizerName, err)
+		return &genericRequeueableError{"failed to add finalizer " + k8sv1.FinalizerName, err}
 	}
 
 	logger.Info("Install helm chart...")
 
-	// create a new context that does not get cancelled immediately on SIGTERM
+	// create a new context that does not get canceled immediately on SIGTERM
 	helmCtx := context.Background()
 
 	if err := cim.helmClient.InstallOrUpgrade(helmCtx, component.GetHelmChartSpec()); err != nil {
-		return fmt.Errorf("failed to install chart for component %s: %w", component.Spec.Name, err)
+		return &genericRequeueableError{"failed to install chart for component " + component.Spec.Name, err}
 	}
 
 	component, err = cim.componentClient.UpdateStatusInstalled(helmCtx, component)
 	if err != nil {
-		return fmt.Errorf("failed to update status-installed for component %s: %w", component.Spec.Name, err)
+		return &genericRequeueableError{"failed to update status-installed for component " + component.Spec.Name, err}
 	}
 
 	logger.Info(fmt.Sprintf("Installed component %s.", component.Spec.Name))